+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - The PythonServer can launch, in the foreground or in the background, Python functions on any folders.
 - It is now possible to specify more than one proxy per service (in case both http and https schemes are required). It is also possible to specify a no_proxy url as well.
 - Authentication is now mostly handled in a separate `requests_auth` module, this module was refactored to get rid of the `flask` dependency.
+- ClickOnce application cache will now be cleared after Microsoft Excel add-in uninstallation to ensure that next installation succeed. Installing add-in will now be possible even if Microsoft Excel is launched.
 
 ### Bug fixes ###
 
@@ -25,6 +26,7 @@
 - PyxelRest is now properly set as a python package and can be used outside Microsoft Excel.
 - file:// swagger URL are now properly analyzed by Microsoft Excel Add-In
 - Only HTTP GET results can now be cached as it does not make sense for other HTTP methods.
+- Do not prompt anymore for one update per closed Microsoft Excel instance. Only prompt the user once.
 
 ## 0.62 (2017-06-27) ##